[LI-CHERRY-PICK] [e4215c1] KAFKA-8325, KAFKA-8202; Remove batch from in-flight requests on MESSAGE_TOO_LARGE errors (#7176)
TICKET = KAFKA-8052, KAFKA-8202
LI_DESCRIPTION =
This patch fixes memory leaks in producer when batch split happens.

EXIT_CRITERIA = HASH [e4215c1]
ORIGINAL_DESCRIPTION =

This patch fixes a bug in the handling of MESSAGE_TOO_LARGE errors. The large batch is split, the smaller batches are re-added to the accumulator, and the batch is deallocated, but it was not removed from the list of in-flight batches. When the batch was eventually expired from the in-flight batches, the producer would try to deallocate it a second time, causing an error. This patch changes the behavior to correctly remove the batch from the list of in-flight requests.

Reviewers: Luke Stephenson, Ismael Juma <ismael@juma.me.uk>, Jason Gustafson <jason@confluent.io>