[LI-HOTFIX] Add a logline to include connection ID when describing a request and response pair
TICKET =
LI_DESCRIPTION =
The method Observer.describeRequestAndResponse is used to generate
request/response information to log when the observer throws an
exception. The generated request/response information will be logged
together with the exception thrown by the observer.
Logging the connection ID along with the exception will help us
understand how the value of connection ID plays a part in the exception
and help us root cause the problem.

EXIT_CRITERIA = Never since this is a linkedin internal change