Add topic availability safety check to controlled shutdown
The controller will not allow a broker to initiate controlled shutdown
unless the shutdown is unlikely to lead to a topic-partition ISR
shrinking below the min.insync.replicas setting for that topic.

This new functionality is optional and disabled by default. The
following new configs have been added:

* controlled.shutdown.safety.check.enable: Enable this safety check
  for any ControlledShutdownRequest handled by the controller.
* controlled.shutdown.topic.cache.size: In order to implement this
  safety check the controller needs access to topic
  configuration. Looking this up requires a read from ZooKeeper, and
  these configs change infrequently, so caching this will vastly
  reduce the amount of time required to perform this safety check on
  large clusters with tens of thousands of topics.
* controlled.shutdown.topic.config.cache.ttl.ms: The number of
  milliseconds to cache topic configs for this safety check. Default
  is one minute.

By default a broker attempting shutdown will only send
ControlledShutdownRequest to the controller 3 times with a 5 second
sleep in between. If it is desired to never skip controlled shutdown,
controlled.shutdown.max.retries can be increased to a much larger
value than 3. It can be set to 2147483647 (Integer.MAX_VALUE) to
effectively retry forever.

A controller integration test has been added which exercises this code
with a simulated three node cluster with a topic configured with
min.insync.replicas=2, allowing only one broker to shut down at a
time.

org.apache.kafka.common.cache.BoundedTimedCache was added to implement
a time-based bounded cache for topic configuration in the controller.

In ControllerContext, instead of maintaining a list of broker ids that
have begun shutting down called shuttingDownBrokers, a map of brokerId
-> brokerEpoch. This will make it easier to keep this map and broker
presence notifications (via zookeeper) in sync.

When a ControlledShutdown request is allowed to begin, it is recorded
in zookeeper under /brokers/shutdown/<broker-id>, with the value of
the broker epoch. On controller initialization this is loaded into
controllerContext.shuttingDownBrokers. When a broker has restarted, it
will come back with a new broker epoch which will take precedence over
the epoch in controllerContext.shuttingDownBrokers.