Add the ability to skip the shutdown safety check
This patch adds LiControlledShutdownSkipSafetyCheckRequest. It takes a
brokerId and brokerEpoch. When sent to the controller, the controller
will record it in a map in memory.

After handling that request, the controller will ignore the results of
the safety check for any request with that brokerId and epoch <= the
epoch provided via LiControlledShutdownSkipSafetyCheckRequest.

This is not persisted to zookeeper since this is an emergency
action. If the controller moves then an admin can just send the
request again.

This request is authorized as a cluster action.

Since LiControlledShutdownSkipSafetyCheck is not submitted to
upstream, it has been added with an ApiKey of 1000. The largest ApiKey
in upstream is 47. This new API is only meant to be used by
AdminClient and is not meant to be an inter-broker request. So this
sets the precedent that if we have an admin-only API we want to add
that isn't upstream yet, we can give it an ApiKey of 1000+. Then if
this internal-only admin API does make it upstream, migrating our
tools to it will be easy since usage should be minimal.

This required fixing a couple of tests that made assumptions about
ApiKeys, namely that the array of api keys now has gaps in it. Now
ApiKeys#hasKeys returns true if the key is within range *and* the
entry in the ID_TO_TYPE array is not null.