[LI-HOTFIX] Add safety check to controlled shutdown (#103)
The controller will not allow a broker to initiate controlled shutdown
unless the shutdown is unlikely to lead to a topic-partition ISR
shrinking below the min.insync.replicas setting for that topic.

This new functionality is optional and disabled by default. The
following new configs have been added:

* controlled.shutdown.safety.check.enable: Enable this safety check
  for any ControlledShutdownRequest handled by the controller.
* controlled.shutdown.safety.check.redundancy.factor: Additional
  replicas above min.insync.replicas that are required for any
  topic-partition to allow shutdown of a broker. Default is 2, so if a
  topic has min.insync.replicas=2, then the ISR must have 4 replicas
  to allow shutdown.

By default a broker attempting shutdown will only send
ControlledShutdownRequest to the controller 3 times with a 5 second
sleep in between. If it is desired to never skip controlled shutdown,
controlled.shutdown.max.retries can be increased to a much larger
value than 3. It can be set to 2147483647 (Integer.MAX_VALUE) to
effectively retry forever.

A controller integration test has been added which exercises this code
with a simulated three node cluster with a topic configured with
min.insync.replicas=2, allowing only one broker to shut down at a
time.

In ControllerContext, instead of maintaining a list of broker ids that
have begun shutting down called shuttingDownBrokers, a map of brokerId
-> brokerEpoch. This will make it easier to keep this map and broker
presence notifications (via zookeeper) in sync.

When a ControlledShutdown request is allowed to begin, it is recorded
in zookeeper under /brokers/shutdown/<broker-id>, with the value of
the broker epoch. On controller initialization this is loaded into
controllerContext.shuttingDownBrokers. When a broker has restarted, it
will come back with a new broker epoch which will take precedence over
the epoch in controllerContext.shuttingDownBrokers.

The controller now needs to know the min.insync.replicas of every
topic. This is done by a combination of DynamicConfigManager,
a full scan of all topic configs on controller failover, and a /topics
znode watcher. All update a map in ControllerContext.

In order to allow skipping the safety check, this patch adds
LiControlledShutdownSkipSafetyCheckRequest. It takes a brokerId
and brokerEpoch. When sent to the controller, the controller
will record it in a map in memory.

After handling that request, the controller will ignore the results of
the safety check for any request with that brokerId and epoch <= the
epoch provided via LiControlledShutdownSkipSafetyCheckRequest.

This is not persisted to zookeeper since this is an emergency
action. If the controller moves then an admin can just send the
request again.

This request is authorized as a cluster action.

Since LiControlledShutdownSkipSafetyCheck is not submitted to
upstream, it has been added with an ApiKey of 1000. The largest ApiKey
in upstream is 47. This new API is only meant to be used by
AdminClient and is not meant to be an inter-broker request. So this
sets the precedent that if we have an admin-only API we want to add
that isn't upstream yet, we can give it an ApiKey of 1000+. Then if
this internal-only admin API does make it upstream, migrating our
tools to it will be easy since usage should be minimal.

This required fixing a couple of tests that made assumptions about
ApiKeys, namely that the array of api keys now has gaps in it. Now
ApiKeys#hasKeys returns true if the key is within range *and* the
entry in the ID_TO_TYPE array is not null.