[LI-HOTFIX] Cover ControllerContext.topicMinIsrConfig map in tests
This is a followup to c337e4752, which added a safety check to the
controlled shutdown protocol. See github.com/linkedin/kafka/pulls/103

This just updates ControllerIntegrationTest case
testControlledShutdownRejectRequestForAvailabilityRisk so that it
covers the logic which was added to maintain the topicMinIsrConfig map
in ControllerContext.

This ensures that the map is updated on topic creation, on controller
failover, and on configuration update.

TICKET = N/A
LI_DESCRIPTION = LIKAFKA-32788
EXIT_CRITERIA = N/A