Feedback from Lucas
He pointed out that if a broker which is attempting to shutdown is
already out of ISR, then it is probably pretty harmless to allow it to
shut down if the safety check would have allowed it to when there are
no under-replicated partitions.

So this changes the logic slightly.

ControlledShutdownSafetyCheckEnableProp is now defined as, "Shutdown
will be allowed if removing the candidate broker will leave
min.insync.replicas + ControlledShutdownSafetyCheckEnableProp replicas
in the ISR.

The default redundancy factor is changed to 1.