[LI-HOTFIX] Skip shutdown safety check at epoch already checked (#157)
TICKET = N/A
LI_DESCRIPTION =
This makes a couple of changes to address some deficiencies
discovered with this safety check when using it in a large, busy
cluster.

* Skip the check if already approved for the requested epoch
* Define redundancy factor property as the required result of
  a broker shutdown, not the required state before shutdown,
  allowing the shutdown of lagging replicas.
* Change default redundancy factor from 2 => 1

In c337e4752 (PR #103) a shutdown safety check was introduced. In
testing this in a large cluster, we found a problem. If the controller
sees that brokerId 1 is safe to shut down at brokerEpoch 99, it will
add that broker to `controllerContext.shuttingDownBrokerIds`. But it
might still not allow shutdown because some leadership transitions
must happen first.

In the process of performing those leadership transitions, brokerId 1
may receive StopReplica, which would cause it to leave the
ISR. Subsequent ControlledShutdownRequest attempts would then fail the
`safeToShutdown` safety check, so the broker would never be allowed to
shut down.

This patch adjusts the implementation of `safeToShutdown`
slightly. Now, if the controller has said that a broker is safe to
shut down at a given epoch, it will always allow that shutdown for
that epoch without performing any additional checks. This allows the
full controlled shutdown process to complete which in a large, busy
cluster may take several rounds of `ControlledShutdownRequest`.

In addition, Lucas pointed out that if a broker which is attempting to
shutdown is already out of ISR, then it is probably pretty harmless to
allow it to shut down if the safety check would have allowed it to when
there are no under-replicated partitions. So we changed the logic a
little further.

ControlledShutdownSafetyCheckEnableProp is now defined as,
 "Shutdown will be allowed if removing the candidate broker will leave
min.insync.replicas + ControlledShutdownSafetyCheckEnableProp replicas
in the ISR."

The default redundancy factor is changed to 1.

EXIT_CRITERIA = N/A