Remote fetch should be enabled only for topics with remote storage enabled.
Summary: In a cluster, there can exist topics with and without remote storage. If a topic is configured with only local storage and if the consumer throws OffsetOutOfRangeException, we shouldn't redirect the fetch request to the remote tier.

Test Plan: Will add unit test once the patch gets deployed in the phx3-ssd1 test cluster.

Reviewers: prateeka, satishd

Reviewed By: satishd

JIRA Issues: DKAFC-947

Differential Revision: https://code.uberinternal.com/D6380461