[LI-HOTFIX] Ensure that the controller has fresh values for min.insync.replicas after topic-level configs are dropped.
TICKET = CM-79574
LI_DESCRIPTION =
The following series of events causes the Controller to keep stale config value for min.insync.replicas config.
This can cause a broker shutdown to be blocked when ISR-based shutdown logic is adopted.

1. Given a topic T, set its topic-level min.insync.replicas config value to XXX.
2. Then remove the topic-level min.insync.replicas config value of topic T

Currently, step-1 will cause the controller update its "controllerContext.topicMinIsrConfig" value for the topic T to XXX.
However, step-2 will not remove the config value XXX from "controllerContext.topicMinIsrConfig". As a result, Controller
will cache a stale config value of XXX for topic T. The expected behavior was for the topic to use the default config value
after its topic-level config was removed.

* The above events can cause the ISR-based shutdown logic to block shutdown. For example,
	1. Set the min.insync.replicas=2 for topic T.
	2. Remvoe topic-level min.insync.replicas config value for topic T.
	3. Use redundancy factor of 1 for ISR-based shutdown logic.
	4. Observe that controller will block shutdown for a partition of the topic with the following log:

	```
	T-2 has min.insync.replicas=2 and a redundancy factor of 1. Removing broker XXX will leave 2 live replicas in the ISR.
	```

EXIT_CRITERIA = When this change is merged in upstream and the corresponding change is pulled into a future release.