[LI-HOTFIX] Add the move controller RPC
This is a squash of 2 commits in 2.4-li
- 782005008e [LI-HOTFIX] Add the move controller RPC (#211)
- 8c0345aef7 [LI-HOTFIX] Fixing the broken tests in RequestQuotaTest caused by the new moveController API (#212)

TICKET = LIKAFKA-37769
LI_DESCRIPTION =
In order to make Kafka SOX compliant, all of the znodes need to be protected with ACLS and become writable only by kafka-server. This means the ktool cluster move-controller command will be rejected.

To continue supporting the ktool command above, this PR adds a new moveController API to the KafkaAdminClient.
EXIT_CRITERIA = We can deprecate this PR if either of the following two conditions is met
1. this change is merged in upstream
2. a new equivalent API is added upstream that works based on the Raft-based controller quorum, and we are deprecating zookeeper.