[LI-HOTFIX] Fixing the broken tests in RequestQuotaTest caused by the new moveController API
TICKET = N/A
LI_DESCRIPTION =
Recently we added the moveController API to the KafkaAdminClient,
which caused some tests in the RequestQuotaTest to fail.
This PR fixes the broken tests by recognizing the new LI_MOVE_CONTROLLER API key.

EXIT_CRITERIA = this change has the same exit criteria as the
original PR that addes the moveController API.