Moved highest remote offset fetching to RLMTask::run() method.
The leader RLMTask needs to know the highest offset that is consistent
with the leader's epoch history and is already copied to the remote tier.
This is done by calling the findHighestRemoteOffset method inside the
RLMTask::convertToLeader() until now. This call can fail if the RLMM is
still initializing, and is not yet ready to take requests. In this scenario,
findHighestRemoteOffset() throws an exception, causing the LeaderAndIsrRequest
to fail and not be processed.

After this change, the method will be called inside an execution of the
RLMTask, and if an exception occurs, it will be retried again in the
next execution. The task will succeed eventually once the RLMM is available
to respond to thehighestOffsetForEpoch method.