[LI-HOTFIX] Honor the result of whether there should be a backoff for TruncateAndFetch
TICKET = N/A

LI_DESCRIPTION =
The current implementation has a bug in that the boolean value returned by
processFetchRequest is discarded. The result indicates whether there should be
a backoff before the next TruncateAndFetch event is enqueued.

This PR addresses this bug by propagating the boolean value back into the
truncateAndFetch method.

EXIT_CRITERIA = N/A