Load pending partitions into metadata store during RLMM init.
Currently, when onPartitionLeadershipChanges is called before initialzation, it adds the partitions to a pendingAssignPartitions set. Then, when TopicBasedRemoteLogMetadataManager gets initialized, it adds the partitions in the pendingAssignPartitions set to the ConsumerManager to consume from, but does not store them in the remotePartitionMetadataStore.

This behavior is different from when `addRemoteLogSegmentMetadata` is called after initialization, and it calls remotePartitionMetadataStore.maybeLoadPartition(partition).

This change makes the behavior consistent, so that the RemotePartitionMetadataStore does not miss any partitions that were added before initialization.