[LI-HOTFIX] Revert "[LI-HOTFIX] Consolidate the interactions between the RequestHandler threads and fetcher threads (#231)"
This reverts commit a34a489ad4509202206b1c75c00678acb4e9a809.
During certification, we found that this change is not safe.
For example, suppose the follower 0 is currently fetching data for partition tp0 from leader 1,
and we are trying to switch the leadership to leader 2.

In the previous implementation, tp0 goes through the following steps
1. tp0 is removed from the fetcher targeting leader 1,
2. retrieving the fetch offset based on the new leader's epoch. Say leader 2 returns a fetch offset of 100.
3. tp0 is added to a new fetcher thread with the fetch offset of 100.

After the PR, the following race condition may happen
1. tp0 retrieves the fetch offset, e.g. 100, from leader 2
2. tp0 is concurrently sent for removal from the fetcher-with-leader-1
   and addition to the fetcher-with-leader-2 with a fetch offset of 100.

Before the removal from fetcher-with-leader-1 is processed, more data may
possible be retrieved from leader 1 and appended to the log (e.g. making the log end offset to become 102).
Then the fetch offset, i.e. 100, would not match the log end offset, i.e. 102.