[LI-HOTFIX] Cache the UpdateMetadataRequest.Builder when the IBP is >= KAFKA_2_3_IV2
TICKET = KAFKA-7186
LI_DESCERIPTION =
The original PR for this change is #75.
Before this patch, the controller needs to allocate an individual UpdateMetadataRequest.Builder
for each broker. This may incur significant memory overhead if the controller has just started,
and is trying to send the cluster's full metadata to all brokers.

This PR tries to reduce the memory footprint by reusing the same UpdateMetadataRequest.Builder.
This is only achievable when all of the UMR have the same body payload, which means
the common UMR should use the maxBrokerEpoch field instead of individual broker epochs.

EXIT_CRITERIA = KAFKA-7186