[LI-HOTFIX] Populate the metadata cache before processing LeaderAndIsr
TICKET = N/A
LI_DESCRIPTION =
When processing a LiCombinedControl request, we found that sometimes the LeaderAndIsr section
depends on the UpdateMetadata section, which implies that the UpdateMetadataRequest section
should be processed before the LeaderAndIsr section.

If the order is reversed, a broker will fail to become
the follower, and result in the following error message
`
Received LeaderAndIsrRequest with correlation id ... from controller ... epoch ...
for partition ... (last update controller epoch ...) but cannot become follower
since the new leader ... is unavailable.
`

EXIT_CRITERIA =