[LI-HOTFIX] Downgrade the CI build to Java 8
TICKET = LIKAFKA-43889
LI_DESCRIPTION =
1. scala compiler run with java8 produces java8 compatible code before 2.4.1.42 (2.4.1.47 also worked since I built it via CLI on my localhost which also used java8).
2. scala compiler run with java11 produces java8 incompatible code. Even though java11 is given the
--release 8 option, the scala compiler is not aware of it. When I tried to add the -release 8 option
to the scala compiler, it throws out this error
`
/home/lucas/workspace/likafka-testing-java11/core/src/main/scala/kafka/Kafka.scala:80: Class sun.reflect.CallerSensitive not found - continuing with a stub.
 Runtime.getRuntime().addShutdownHook(new Thread("kafka-shutdown-hook") {
 ^
`

It seems we have the following options # Chase the error above and try to fix it

1. Upgrade samza-li to java 11
2. Downgrade the kafka CI environment to java 8.
At this moment, the Samza team says option 1 is not acceptable.
Thus we are left with option 2.

EXIT_CRITERIA =
In the future, when samza-li is upgraded to java 11 or when it no longer depends on
the kafka scala code, we can do the upgrade to java 11 again.