[LI-HOTFIX] Improve the efficiency of replica deletion
TICKET = TBD
LI_DESCRIPTION =
KAFKA-10002 identified that deletion of replicas can be slow when a StopReplica request is being
processed, and has implemented a change to improve the efficiency.
We found that the efficiency can be further improved by avoiding the reinitialization of the
leader epoch cache and partition metadata for a replica that needs to be deleted.

EXIT_CRITERIA = TBD