[LI-HOTFIX] Resume topic deletion on broker failure
TICKET = KAFKA-10548
LI_DESCRIPTION =
In the current implementation, a topic goes through the following phases during deletion
1. the topic is added to the ControllerContext.topicsToBeDeleted
2. replicas of the topic are moved to ReplicaDeletionStarted if the broker is online, or
   to ReplicaDeletionSuccessful if the broker is offline
3. upon receiving a response of the StopReplica request from the broker, a replica's state is
   updated to ReplicaDeletionSuccessful, or to ReplicaDeletionIneligible if there is
   an error in the response

However, it does not cover the case where a broker immediately fails after a topic goes through
the above steps for deletion. Under such conditions, the replicas for the to-be-deleted topic
would stay in ReplicaDeletionStarted, and the deletion would still be stuck.

This PR fixes the issue by resuming the deletion of a topic upon broker failure.
It makes the following changes
1. The deletion success criteria has been changed from "all replicas in ReplicaDeletionSuccessful"
   to "all replicas in ReplicaDeletionSuccessful or OfflineReplica".
2. During deletion, an offline broker's replicas will stay in the state OfflineReplica,
   and will no longer be moved to ReplicaDeletionSuccessful. This change is for easier
   understanding.
3. Upon broker failure, the topics of the new offline replicas is checked for deletion,
   and will have their deletion resumed if eligible.

EXIT_CRITERIA = When changes in KAFKA-10548 merged in upstream and the changes are pulled in