[LI-FIXUP] Fix dynamic delete.topic.enable flag not handled in broker request code path
This should be a direct fix-up to the commit:
- b5fa0af [LI-HOTFIX] Add topic deletion hook in zookeeper to trigger in-memory topic deletion flag changes

TICKET =
LI_DESCRIPTION =
When the ZK hook was introduced, it didn't change the KafkaApi handler to make it honor the internal state.
Therefore, the deletion would still be blocked unless it's directly writing to the `/admin/delete_topics` ZNode.
This patch fixes the code in API handling by change the reference to `config.deleteTopicEnabled` to `TopicDeletionManager`'s  internal state.

EXIT_CRITERIA = "When the original dynamic flag commit is ejected"