[LI-HOTFIX] Keep LiCombinedControl request enabled when IBP<2.8
TICKET = N/A
LI_DESCRIPTION =
The upgrade from kafka 2.4 to 3.0 is a 2-step process.
Code 2.4/IBP=2.4
---1st upgrade-->
Code 3.0/IBP=2.4
---2nd upgrade-->
Code 3.0/IBP=3.0

Considering the Code 3.0/IBP=2.4 state in the middle, the current implementation requires
that the first full LeaderAndIsr request to be sent before the LiCombinedControl request
feature can be enabled. However, the full LeaderAndIsr is only enabled when the IBP>=2.8.
It means the LiCombinedControl request feature is disabled in the interim state, which
may cause delays in metadata propagation and further cause GCNs.

This patch ensures that the LiCombinedControl feature is not disabled during the upgrade.

EXIT_CRITERIA = The same with the LiCombinedControl feature