[LI-HOTFIX] Add metric to track ProduceRequestsWithInvalidAcksPerSec
TICKET = LIKAFKA-45529
LI_DESCRIPTION =
Inside LI, if a topic has minISR=2, we'd expect all of the topic's corresponding producers
should have the durable acks setting, i.e. acks=-1.

Currently, there is no metric to track whether the producers' acks settings are correct.
This PR adds the metric, so that we can discover the producers with invalid acks settings.

EXIT_CRITERIA = N/A as this policy is specific to LI.