[LI-HOTFIX] Add extra logging to check if the Selecter is notified of incoming data
TICKET = GCN-37194
LI_DESCRIPTION =
In the GCN, we are seeing that the client timed out while waiting for a MetadataResponse to come.
However, wireshark packet capture seems to suggest that the response did arrive for the client.

This PR adds extra logging to check whether the Selector is notified of the incoming data
and is able to read from the socket.
EXIT_CRITERIA = N/A