[LI-HOTFIX] Enable LiCombinedControl for nondata brokers
TICKET = LIKAFKA-46387

LI_DESCRIPTION = In the current implementation, the LiCombinedControl requests are enabled only
after a full LeaderAndIsr request is sent. For the preferred controllers and maintenance brokers,
they'll never receive a LeaderAndIsr request, and thus have the LiCombinedControl requests
permanently disabled.

This PR fixes the problem by enabling the LiCombinedControl requests for the preferred
controllers and maintenance brokers.

EXIT_CRITERIA = The same criteria as the LiCombinedControl feature.