[LI-HOTFIX] Add the li.alter.isr.enable config and use the ZK based AlterIsrManager by default
TICKET = LIKAFKA-46520
LI_DESCRIPTION =
We are seeing the AlterISR request can take a long time in the extreme condiditions.

And when a leader has an inflight AlterISR request, the subsequent AlterISR requests are dropped,
indicated by the following log from the leader:

"Failed to enqueue ISR change state LeaderAndIsr(leader=..., leaderEpoch=.., isr=List(...),
zkVersion=...) for partition ..."
This means the followers are not able to join the ISR quickly and hence cause UnderMinISR partitions
to last for a long time.

EXIT_CRITERIA = When we make sure the AlterISR requests are efficient enough for our use case.