[LI-HOTFIX] Enable LiCombinedControl for brokers without any replicas
TICKET = N/A
LI_DESCRIPTION =
Currently the LiCombinedControl requests are explicitly enabled for the preferred controllers
and maintenance brokers. This PR improves the condition so that all brokers that don't
have any replicas assigned will have the LiCombinedControl request enabled.

To achieve this, we add a new variable inside the ControllerContext called
lazilyUpdatedAssignmentsByBroker, whose source of truth is the partitionAssignments variable.
lazilyUpdatedAssignmentsByBroker is not updated each time there is a change in partitionAssignments
since doing that may significantly slow down topic creation and deletion.
Instead, it's only lazily updated when necessary, i.e. when querying per broker assignments.

EXIT_CRITERIA = N/A