[LI-HOTFIX] Improve the DefaultAlterIsrManager and use it by default
TICKET = LIKAFKA-46520
LI_DESCRIPTION =
In PR #386, we switched to use the ZkIsrManager by default, since we were seeing the following error
after the 3.0 upgrade.
"Failed to enqueue ISR change state LeaderAndIsr(leader=..., leaderEpoch=.., isr=List(...),
zkVersion=...) for partition ..."

For unknown reasons, the request to update ZK isr was blocked in several clusters when the change
was rolled out.

This PR tries to improve the DefaultAlterIsrManager and switch it back as the default.
Specifically, we add a new unsentIsrQueue such that submission of an AlterIsrItem would always succeed.
EXIT_CRITERIA = When this changed is submitted upstream and pulled back